test(table): add a test for suppressed vulns
