+++ pkg/report/table/vulnerability.go
@@ -175,10 +175,10 @@ func (r *vulnerabilityRenderer) renderModifiedVulnerabilities() {
 
 	var total int
 	for _, m := range r.result.ModifiedFindings {
-		vuln, ok := m.Finding.(types.DetectedVulnerability)
-		if !ok {
+		if m.Type != types.FindingTypeVulnerability {
 			continue
 		}
+		vuln := m.Finding.(types.DetectedVulnerability)
 		total++
 
 		stmt := lo.Ternary(m.Statement != "", m.Statement, "N/A")
